refactor(expansion): decouple wildcard logic from main expansion loop
- Delete `add_in_wildcard_hashmap.c` to remove the now-obsolete hashmap-based approach for tracking wildcard characters.
- In `expand_dollar_word.c`, modify `add_word_in_arg_list` to process the generated word through `expand_wildcard` and then merge the results using `join_list`.
- In `expand_regular_word.c`, simplify the call to the `utils` function by removing the now-unnecessary boolean flag for wildcard handling.
- In `expand_single_quote_word.c`, update the call to `utils` by removing the redundant boolean parameter previously used for wildcard logic.
- Introduce `expand_wildcard.c` with a stub function to encapsulate the new, simplified wildcard expansion logic that operates on a complete word pattern.
- In `expansion.c`, refactor the `expander` function to remove the wildcard hashmap and incorporate the new expansion flow using `expand_wildcard` and `join_list`.
- In `expansion_utils.c`, simplify the `utils` function signature by removing the `wildcard_falg` parameter, as wildcard tracking is no longer handled at this stage.
- In `expansion.h`, update the `t_info` struct by removing the `hashmap` field and adjust function prototypes to align with the refactored expansion logic.
- In `nsh.h`, change the data type of the `size` member in `s_list_info` from `int` to `size_t` for better type correctness and consistency.
- In `utils.h`, add the function prototype for the new `join_list` helper to make it available for use in other parts of the codebase.
- In `join_list.c`, add the new utility function to efficiently concatenate two `t_list_info` linked lists.
- In `join_two_strings.c`, fix a critical bug by adding the missing `return` statement to ensure the function returns the concatenated string.

+++ includes/nsh.h
@@ -6,7 +6,7 @@
 /*   By: yjaafar <yjaafar@student.1337.ma>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/01 08:40:39 by yjaafar           #+#    #+#             */
-/*   Updated: 2025/08/07 02:11:32 by yjaafar          ###   ########.fr       */
+/*   Updated: 2025/08/07 05:03:03 by yjaafar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,7 +38,7 @@ typedef struct s_list_info
 {
 	t_list		*list;
 	t_list		*tail;
-	int			size;
+	size_t		size;
 }	t_list_info;
 
 typedef struct s_token
